+++ crates/acts/src/scheduler/process/task/step.rs
@@ -35,7 +35,7 @@ impl ActTask for Step {
         Ok(())
     }
 
-    fn next(&self, ctx: &Context) -> Result<NextAction> {
+    async fn next(&self, ctx: &Context) -> Result<NextAction> {
         let task = ctx.task();
 
         if task.state().is_success() || task.state().is_skip() {