+++ crates/acts/src/scheduler/tests/mod.rs
@@ -1,4 +1,5 @@
 mod act;
+mod action;
 mod branch;
 mod message;
 mod proc;